[T] move: move component
