+++ __tests__/containers/discover/roadmap-list/roadmapList.test.js
@@ -2,7 +2,7 @@ import 'react-native';
 import React, {} from 'react';
 import renderer from 'react-test-renderer';
 import RoadmapList from '../../../../src/containers/discover/roadmap-list/RoadmapList';
-import TIPS from '../../../../assets/TIPS';
+import TIPS from '../../../../assets/data/TIPS';
 
 it('renders correctly', () => {
   const roadmapList = renderer.create(
