+++ __tests__/containers/discover/toolbox-list/toolBoxList.test.js
@@ -2,7 +2,7 @@ import 'react-native';
 import React, {} from 'react';
 import renderer from 'react-test-renderer';
 import ToolBoxList from '../../../../src/containers/discover/toolbox-list/ToolBoxList';
-import TIPS from '../../../../assets/TIPS';
+import TIPS from '../../../../assets/data/TIPS';
 
 it('renders correctly', () => {
   const toolBoxList = renderer.create(
