+++ src/components/discover/view/DiscoverOnlineReadItem.js
@@ -1,8 +1,8 @@
 import React, { Component, PropTypes } from 'react';
 import { View, Text, Image, TouchableHighlight } from 'react-native';
-import AppStyle from '../../theme/styles';
+import AppStyle from '../../../theme/styles';
 
-class MoreItem extends Component {
+class OnLineReadItem extends Component {
   static propTypes = {
     imageParentStyle: View.propTypes.style,
     image: PropTypes.oneOfType([
@@ -20,19 +20,17 @@ class MoreItem extends Component {
         }))]),
     imageStyle: Image.propTypes.style,
     title: PropTypes.string.isRequired,
-    titleColor: PropTypes.string,
     titleStyle: Text.propTypes.style,
     onclick: PropTypes.func,
     top: PropTypes.number,
   };
 
   static defaultProps = {
-    imageParentStyle: AppStyle.discoverMoreItem,
+    imageParentStyle: AppStyle.discoverOnlineItem,
     image: '',
-    imageStyle: AppStyle.discoverMoreItemImage,
+    imageStyle: AppStyle.discoverOnlineItemImage,
     title: '',
-    titleColor: '#03a9f4',
-    titleStyle: AppStyle.discoverMoreItemTitle,
+    titleStyle: AppStyle.discoverOnlineItemTitle,
     onclick: () => {},
     top: 0,
   };
@@ -43,13 +41,12 @@ class MoreItem extends Component {
       style={{ flex: 1, marginTop: this.props.top }}
     >
       <View style={this.props.imageParentStyle}>
-        <Image source={this.props.image} style={this.props.imageStyle} >
-          <View style={this.props.titleStyle}>
-            <Text style={{ color: this.props.titleColor }}>{this.props.title}</Text>
-          </View>
-        </Image>
+        <Image source={this.props.image} style={this.props.imageStyle} />
+        <View style={this.props.titleStyle}>
+          <Text>{this.props.title}</Text>
+        </View>
       </View>
     </TouchableHighlight>
   );
 }
-export default MoreItem;
+export default OnLineReadItem;
