+++ src/components/discover/view/ProjectListItem.js
