+++ src/components/discover/view/TimeLineView.js
@@ -1,8 +1,8 @@
 import React, { Component, PropTypes } from 'react';
 import { View, Text } from 'react-native';
-import Icon from '../../icon/GrowthFont';
+import Icon from '../../../icon/GrowthFont';
 
-import AppStyle from '../../theme/styles';
+import AppStyle from '../../../theme/styles';
 
 class TimeLine extends Component {
   static componentName = 'TimeLine';
